+++ src/Layers/BetterMDPopup.cpp
@@ -16,13 +16,24 @@ bool BetterMDPopup::init(FLAlertLayerProtocol* delegate, char const* title, gd::
 
     auto content = public_cast(md, m_content);
 
+    std::unordered_map<float, std::vector<CCNode*>> posMap;
+    std::unordered_map<float, float> widthMap;
+
     for (auto child : CCArrayExt<CCNode*>(content->getChildren()))
     {
-        //child->setPositionX(child->getPositionX() + content->getContentWidth() / 2);
+        posMap[child->getPositionY()].push_back(child);
+        widthMap[child->getPositionY()] += child->getScaledContentWidth();
+    }
+
+    for (auto obj : posMap)
+    {
+        for (auto child : obj.second)
+        {
+            child->setPositionX(md->getContentWidth() / 2 + child->getPositionX() - widthMap[obj.first] / 2);
+        }
     }
 
     m_mainLayer->addChild(md, 69);
-
     handleTouchPriority(this);
 
     return true;
@@ -40,4 +51,9 @@ BetterMDPopup* BetterMDPopup::create(FLAlertLayerProtocol* delegate, char const*
 
     CC_SAFE_DELETE(pRet);
     return nullptr;
+}
+
+BetterMDPopup* BetterMDPopup::create(char const* title, gd::string desc, char const* btn1, char const* btn2)
+{
+    return BetterMDPopup::create(nullptr, title, desc, btn1, btn2, 420, true, 69, 1.0f);
 }